Fix enums type stubs
