+++ .pre-commit-config.yaml
@@ -119,7 +119,6 @@ repos:
           "--config", "pyproject.toml",
           "--allow-incomplete-defs",
         ]
-        exclude: "nautilus_trader/common/enums.pyi|nautilus_trader/model/enums.pyi"
         additional_dependencies: [
           msgspec,
           types-frozendict,
